use own int/long min and max values; were not available everywhere

+++ animation_nodes/utils/clamp.pxd
@@ -1,3 +1,3 @@
 cdef int clampInt(object value)
-cdef int clampLong(object value)
+cdef long clampLong(object value)
 cdef double clamp(double value, double minValue, double maxValue)
